# docs/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admins/{id}": {
      "get": {
        "operationId": "AdminManagementController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Staff Management"
        ]
      },
      "patch": {
        "operationId": "AdminManagementController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Staff Management"
        ]
      }
    },
...
    "/api/v1/admin-roles": {
      "get": {
        "description": "Admin Roles / RBAC manages permission roles for ADMIN staff users only. SUPER_ADMIN has full immutable access and does not depend on AdminRole permissions.",
        "operationId": "AdminRolesController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isSystem",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Roles / RBAC"
        ]
      },
      "post": {
        "operationId": "AdminRolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Roles / RBAC"
        ]
      }
    },
    "/api/v1/admin-roles/{id}": {
      "get": {
        "operationId": "AdminRolesController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Roles / RBAC"
        ]
      },
      "patch": {
        "operationId": "AdminRolesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Roles / RBAC"
        ]
      },
      "delete": {
        "operationId": "AdminRolesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Roles / RBAC"
        ]
      }
    },
    "/api/v1/admin-roles/{id}/permissions": {
      "patch": {
        "operationId": "AdminRolesController_updatePermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Roles / RBAC"
        ]
      }
    },
    "/api/v1/permissions/catalog": {
      "get": {
        "description": "Read-only list of backend-supported permission keys that can be assigned to admin roles.",
        "operationId": "PermissionCatalogController_catalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Roles / RBAC"
        ]
      }
    },
...
    "/api/v1/broadcasts": {
      "post": {
        "operationId": "BroadcastsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      },
      "get": {
        "operationId": "BroadcastsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "DRAFT",
                "SCHEDULED",
                "PROCESSING",
                "SENT",
                "CANCELLED",
                "FAILED"
              ]
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EMAIL",
                "PUSH",
                "IN_APP"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "NORMAL",
                "HIGH",
                "URGENT"
              ]
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduledFrom",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduledTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "scheduledAt",
                "title",
                "status"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}": {
      "get": {
        "operationId": "BroadcastsController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      },
      "patch": {
        "operationId": "BroadcastsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}/targeting": {
      "patch": {
        "operationId": "BroadcastsController_targeting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastTargetingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/estimate-reach": {
      "post": {
        "operationId": "BroadcastsController_estimate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateReachDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}/schedule": {
      "patch": {
        "operationId": "BroadcastsController_schedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}/cancel": {
      "post": {
        "operationId": "BroadcastsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}/report": {
      "get": {
        "operationId": "BroadcastsController_report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
    "/api/v1/broadcasts/{id}/recipients": {
      "get": {
        "operationId": "BroadcastsController_recipients",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EMAIL",
                "PUSH",
                "IN_APP"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "QUEUED",
                "SENT",
                "DELIVERED",
                "FAILED",
                "OPENED",
                "CLICKED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Broadcast Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gift App Backend API",
    "description": "Gift App authentication and user account APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateAdminRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "UpdateAdminRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "UpdateRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "object"
          }
        },
        "required": [
          "permissions"
        ]
      },
...
      "CreateBroadcastDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "ctaLabel": {
            "type": "string"
          },
          "ctaUrl": {
            "type": "string"
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "EMAIL",
                "PUSH",
                "IN_APP"
              ]
            }
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ]
          }
        },
        "required": [
          "title",
          "message",
          "channels"
        ]
      },
      "UpdateBroadcastDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "ctaLabel": {
            "type": "string"
          },
          "ctaUrl": {
            "type": "string"
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "EMAIL",
                "PUSH",
                "IN_APP"
              ]
            }
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ]
          }
        }
      },
      "TargetingFiltersDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "onlyVerifiedEmails": {
            "type": "boolean"
          },
          "excludeUnsubscribed": {
            "type": "boolean"
          },
          "excludeSuspended": {
            "type": "boolean"
          }
        }
      },
      "BroadcastTargetingDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "ALL_USERS",
              "SPECIFIC_ROLES",
              "CUSTOM_SEGMENT"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ADMIN",
                "PROVIDER",
                "REGISTERED_USER"
              ]
            }
          },
          "filters": {
            "$ref": "#/components/schemas/TargetingFiltersDto"
          }
        },
        "required": [
          "mode"
        ]
      },
      "EstimateReachDto": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "EMAIL",
                "PUSH",
                "IN_APP"
              ]
            }
          },
          "targeting": {
            "$ref": "#/components/schemas/BroadcastTargetingDto"
          }
        },
        "required": [
          "channels",
          "targeting"
        ]
      },
      "ScheduleBroadcastDto": {
        "type": "object",
        "properties": {
          "sendMode": {
            "type": "string",
            "enum": [
              "NOW",
              "SCHEDULED"
            ]
          },
          "scheduledAt": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "isRecurring": {
            "type": "boolean"
          },
          "recurrence": {
            "type": "object"
          }
        },
        "required": [
          "sendMode"
        ]
      },
      "CancelBroadcastDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  }
}